% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-data.struct-addcolumn.r
\name{addcolumn}
\alias{addcolumn}
\title{Converting a vector with a name into a column in the datastruct}
\usage{
\S4method{addcolumn}{data.struct}(x, ...)
}
\description{
Converting a vector with a name into a column in the datastruct
}
