% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-data.struct-as.data.frame.r
\name{as.data.frame.data.struct}
\alias{as.data.frame.data.struct}
\title{Converting custom datastruct to R data.frame}
\usage{
\method{as.data.frame}{data.struct}(x, row.names = NULL, optional = FALSE, ..., cols)
}
\description{
Converting custom datastruct to R data.frame
}
