% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-dose.matrix-as.matrix.r
\name{as.matrix.dose.matrix}
\alias{as.matrix.dose.matrix}
\title{Converting dose matrices to R matrices}
\usage{
\method{as.matrix}{dose.matrix}(x, ...)
}
\description{
Converting dose matrices to R matrices
}
