% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-snp.matrix-as.matrix.r
\name{as.matrix.snp.matrix}
\alias{as.matrix.snp.matrix}
\title{Converting snp matrices to R matrices}
\usage{
\method{as.matrix}{snp.matrix}(x, ...)
}
\description{
Converting snp matrices to R matrices
}
