% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-data.struct-getcolumn.r
\name{$,data.struct-method}
\alias{$,data.struct-method}
\title{Getters for a column in a datastruct, by name}
\usage{
\S4method{$}{data.struct}(x, name)
}
\description{
Getters for a column in a datastruct, by name
}
