% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-4-data.struct.r
\docType{class}
\name{data.struct}
\alias{data.struct}
\title{Data Structure}
\usage{
data.struct(...)
}
\description{
A 'data.frame' like class of object for SNP stats etc.

A constructor for data.struct from a dataframe
}
