% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-dose.matrix-subsetting.r,
%   R/method-snp.matrix-subsetting.r
\name{extract}
\alias{extract}
\alias{[,dose.matrix,missing,logical-method}
\alias{[,dose.matrix,numeric,missing-method}
\alias{[,dose.matrix,logical,missing-method}
\alias{[,dose.matrix,index,index-method}
\alias{[,snp.matrix,missing,logical-method}
\alias{[,snp.matrix,numeric,missing-method}
\alias{[,snp.matrix,logical,missing-method}
\alias{[,snp.matrix,index,index-method}
\title{Extract parts of DOSE matrix}
\usage{
\S4method{[}{dose.matrix,missing,numeric}(x, i, j, ..., drop = TRUE)

\S4method{[}{dose.matrix,missing,logical}(x, i, j, ..., drop = TRUE)

\S4method{[}{dose.matrix,numeric,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{dose.matrix,logical,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{dose.matrix,index,index}(x, i, j, ..., drop = TRUE)

\S4method{[}{snp.matrix,missing,numeric}(x, i, j, ..., drop = TRUE)

\S4method{[}{snp.matrix,missing,logical}(x, i, j, ..., drop = TRUE)

\S4method{[}{snp.matrix,numeric,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{snp.matrix,logical,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{snp.matrix,index,index}(x, i, j, ..., drop = TRUE)
}
\description{
Extract columns / lines of a dose.matrix object

Extract columns / lines of a snp.matrix object
}
