% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.matrix-GRM.r
\name{GRM,snp.matrix-method}
\alias{GRM,snp.matrix-method}
\title{Genetic Relationship Matrix}
\usage{
\S4method{GRM}{snp.matrix}(x, ...)
}
\arguments{
\item{x}{A \code{\link{bed.matrix}}}

\item{...}{extra arguments for some methods. Could be a file name or 
a bool speciifying to use float}
}
\value{
A matrix or mmatrix of GRM values.
}
\description{
Compute the Genetic Relationship Matrix (GRM) between given SNPs.
}
\details{
The GRM is a symmetric square matrix of dimension equal to the number of individuals.
Each entry can be interpreted as an estimated kinship coefficient between individuals,
although some authors might disagree. Note in particular that some entries will be negative.
}
\keyword{Genetic}
\keyword{Matrix}
\keyword{Relationship}
