% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-snp.matrix-GRM.r
\name{GRM}
\alias{GRM}
\title{Genetic Relationship Matrix}
\usage{
GRM(x, ...)
}
\arguments{
\item{x}{A \code{\link{bed.matrix}}}

\item{...}{extra arguments for some methods.}
}
\value{
A matrix or mmatrix.
}
\description{
Compute the Genetic Relationship Matrix (GRM) between given SNPs.
}
\keyword{Genetic}
\keyword{Matrix}
\keyword{Relationship}
