% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.matrix-LD.r
\name{LD}
\alias{LD}
\title{Linkage Disequilibrium}
\usage{
LD(x, lim, lim2, measure, method, ...)
}
\arguments{
\item{x}{A \code{\link{bed.matrix}}}

\item{lim}{Range of SNPs for which the LD is computed}

\item{lim2}{(Optional) Second range of SNPs (see Details)}

\item{measure}{The LD measure}

\item{filename}{if non-missing, the LD matrix will be a mmatrix object}

\item{float}{if TRUE, computations are done in float (only if filename is provided)}
}
\value{
A matrix of LD values.
}
\description{
Compute Linkage Disequilibrium (LD) between given SNPs.
}
\details{
If \code{lim2} is missing, the LD is computed between all SNPs with indices between \code{lim[1]} and \code{lim[2]};
else, the LD is computed between the SNPs in the range given by \code{lim} and those in the range given by \code{lim2}.

The default is to use LD moment estimators (which are faster but less precise than maximum likelihood estimators
obtained with the EM algorithm).
}
\seealso{
\code{\link{LD.thin}},  \code{\link{LD.plot}}
}
\keyword{Disequilibrium}
\keyword{Linkage}
