% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-snp.matrix-LD.thin.r
\name{LD.thin}
\alias{LD.thin}
\title{LD thinning}
\usage{
LD.thin(
  x,
  threshold,
  max.dist,
  beg,
  end,
  which.snps,
  dist.unit,
  extract,
  keep,
  ...
)
}
\description{
LD thinning
}
