% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.r
\name{test.read.bed.matrix}
\alias{test.read.bed.matrix}
\title{test read bed matrix}
\usage{
test.read.bed.matrix(
  basename,
  bed = paste(basename, ".bed", sep = ""),
  fam = paste(basename, ".fam", sep = ""),
  bim = paste(basename, ".bim", sep = "")
)
}
\description{
test read bed matrix
}
\examples{
file <- system.file("extdata", "LCT.bed", package = "snipsnop")
a <- test.read.bed.matrix(file)
all(a == LCT.test)
}
