% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.matrix-binding.r
\name{bind}
\alias{bind}
\alias{rbind2_snp_memory}
\alias{rbind2,snp.matrix}
\alias{cbind2,snp.matrix}
\title{Binding methods for snp.matrix objects}
\usage{
\S4method{rbind2}{snp.matrix,snp.matrix}(x,y,...)

\S4method{cbind2}{snp.matrix,snp.matrix}(x,y,...)
TODO : check if missing version necessary ?

\S4method{cbind2}{snp.matrix,missing}(x,y,...)
}
\arguments{
\item{x, y}{snp.matrix objects}

\item{...}{extra parameters (ignored)}
}
\value{
A (new) snp.matrix matrix combining the arguments.
}
\description{
Methods allowing to use `cbind` and `rbind` with snp.matrix objects.
}
\examples{
x <- dual( c(1, 3) )
y <- cbind(x, 2*x+1, 3*x+2, c(0,1))
y
d(y, "x1")


}
