% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.matrix-subsetting.r
\name{extract}
\alias{extract}
\alias{[,snp.matrix,missing,logical,ANY-method}
\alias{[,snp.matrix,numeric,missing,ANY-method}
\alias{[,snp.matrix,logical,missing,ANY-method}
\alias{[,snp.matrix,index,index,ANY-method}
\title{Extract parts of SNP matrix}
\usage{
\S4method{[}{snp.matrix,missing,numeric,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{snp.matrix,missing,logical,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{snp.matrix,numeric,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{snp.matrix,logical,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{snp.matrix,index,index,ANY}(x, i, j, ..., drop = TRUE)
}
\description{
Extract columns / lines of a snp.matrix object
}
