% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.matrix-GRM.r
\name{GRM,snp.matrix-method}
\alias{GRM,snp.matrix-method}
\title{Genetic Relationship Matrix}
\usage{
\S4method{GRM}{snp.matrix}(x, ...)
}
\arguments{
\item{x}{A \code{\link{bed.matrix}}}

\item{...}{extra arguments for some methods.}
}
\value{
A matrix or mmatrix.
}
\description{
Compute the Genetic Relationship Matrix (GRM) between given SNPs.
}
\keyword{Genetic}
\keyword{Matrix}
\keyword{Relationship}
